+++ changelog.md
@@ -6,6 +6,8 @@
 This version is not released yet and is under active development.
 ```
 
+- Print all modified loggers and their levels in `DEBUG` mode.
+
 ## {gh}`4.0.0 (2023-05-08) <compare/v3.10.0...v4.0.0>`
 
 - Drop support for Python 3.7.